+++ sources/packages/api/src/app.aest.module.ts
@@ -30,7 +30,7 @@ import {
 } from "./route-controllers";
 import { AuthModule } from "./auth/auth.module";
 import { LoggerModule } from "./logger/logger.module";
-import { ApplicationControllerService } from "./route-controllers/application/application.controller";
+import { ApplicationControllerService } from "./route-controllers/application/application.controller.service";
 
 @Module({
   imports: [LoggerModule, DatabaseModule, AuthModule],
