+++ sources/packages/api/src/app.students.module.ts
@@ -25,7 +25,7 @@ import {
 import { ApplicationStudentsController } from "./route-controllers";
 import { AuthModule } from "./auth/auth.module";
 import { LoggerModule } from "./logger/logger.module";
-import { ApplicationControllerService } from "./route-controllers/application/application.controller";
+import { ApplicationControllerService } from "./route-controllers/application/application.controller.service";
 
 @Module({
   imports: [LoggerModule, DatabaseModule, AuthModule],
