+++ sources/packages/api/src/route-controllers/application/models/application.model.ts
@@ -258,58 +258,3 @@ export interface StudentApplicationAndCount {
   applications: ApplicationSummaryDTO[];
   totalApplications: number;
 }
-
-/**
- * Transformation util for Application.
- * @param application
- * @param additionalFormData
- * @returns Application DTO
- */
-export const transformToApplicationDto = (
-  application: Application,
-): GetApplicationBaseDTO => {
-  return {
-    data: application.data,
-    id: application.id,
-    applicationStatus: application.applicationStatus,
-    applicationNumber: application.applicationNumber,
-    applicationFormName: application.programYear.formName,
-    applicationProgramYearID: application.programYear.id,
-  } as GetApplicationBaseDTO;
-};
-
-/**
- * Transformation util for Application.
- * @param applicationDetail
- * @param additionalFormData
- * @returns Application DTO
- */
-export const transformToApplicationDetailDto = (
-  applicationDetail: Application,
-  disbursement: DisbursementSchedule,
-): GetApplicationDataDto => {
-  return {
-    data: applicationDetail.data,
-    id: applicationDetail.id,
-    applicationStatus: applicationDetail.applicationStatus,
-    applicationStatusUpdatedOn: applicationDetail.applicationStatusUpdatedOn,
-    applicationNumber: applicationDetail.applicationNumber,
-    applicationOfferingIntensity: applicationDetail.offering?.offeringIntensity,
-    applicationStartDate: dateString(
-      applicationDetail.offering?.studyStartDate,
-    ),
-    applicationEndDate: dateString(applicationDetail.offering?.studyEndDate),
-    applicationInstitutionName: applicationDetail.location?.name,
-    applicationPIRStatus: applicationDetail.pirStatus,
-    applicationAssessmentStatus: applicationDetail.assessmentStatus,
-    applicationFormName: applicationDetail.programYear.formName,
-    applicationProgramYearID: applicationDetail.programYear.id,
-    applicationPIRDeniedReason: getPIRDeniedReason(applicationDetail),
-    programYearStartDate: applicationDetail.programYear.startDate,
-    programYearEndDate: applicationDetail.programYear.endDate,
-    applicationCOEStatus: disbursement?.coeStatus,
-    applicationCOEDeniedReason: disbursement
-      ? getCOEDeniedReason(disbursement)
-      : undefined,
-  };
-};
